perf: custom fields 仅custom 类型资产支持
