+++ apps/assets/serializers/asset/common.py
@@ -1,18 +1,18 @@
 # -*- coding: utf-8 -*-
 #
 
-import re
-
-from django.db.models import F, QuerySet
+from django.db.models import F
 from django.db.transaction import atomic
 from django.utils.translation import ugettext_lazy as _
 from rest_framework import serializers
 
 from accounts.models import Account
 from accounts.serializers import AccountSerializer
-from common.serializers import WritableNestedModelSerializer, SecretReadableMixin, CommonModelSerializer, \
-    MethodSerializer
-from common.serializers.dynamic import create_serializer_class
+from common.const import UUID_PATTERN
+from common.serializers import (
+    WritableNestedModelSerializer, SecretReadableMixin,
+    CommonModelSerializer, MethodSerializer
+)
 from common.serializers.fields import LabeledChoiceField
 from orgs.mixins.serializers import BulkOrgResourceModelSerializer
 from ...const import Category, AllTypes
@@ -24,8 +24,6 @@ __all__ = [
     'AssetDetailSerializer', 'DetailMixin', 'AssetAccountSerializer',
     'AccountSecretSerializer', 'AssetProtocolsPermsSerializer'
 ]
-
-uuid_pattern = re.compile(r'[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}')
 
 
 class AssetProtocolsSerializer(serializers.ModelSerializer):
@@ -121,12 +119,11 @@ class AssetSerializer(BulkOrgResourceModelSerializer, WritableNestedModelSeriali
     protocols = AssetProtocolsSerializer(many=True, required=False, label=_('Protocols'), default=())
     accounts = AssetAccountSerializer(many=True, required=False, allow_null=True, write_only=True, label=_('Account'))
     nodes_display = serializers.ListField(read_only=False, required=False, label=_("Node path"))
-    custom_info = MethodSerializer(label=_('Custom info'))
 
     class Meta:
         model = Asset
         fields_mini = ['id', 'name', 'address']
-        fields_small = fields_mini + ['custom_info', 'is_active', 'comment']
+        fields_small = fields_mini + ['is_active', 'comment']
         fields_fk = ['domain', 'platform']
         fields_m2m = [
             'nodes', 'labels', 'protocols',
@@ -192,36 +189,6 @@ class AssetSerializer(BulkOrgResourceModelSerializer, WritableNestedModelSeriali
             .annotate(category=F("platform__category")) \
             .annotate(type=F("platform__type"))
         return queryset
-
-    def get_custom_info_serializer(self):
-        request = self.context.get('request')
-        default_field = serializers.DictField(required=False, label=_('Custom info'))
-
-        if not request:
-            return default_field
-
-        if self.instance and isinstance(self.instance, (QuerySet, list)):
-            return default_field
-
-        if not self.instance and uuid_pattern.findall(request.path):
-            pk = uuid_pattern.findall(request.path)[0]
-            self.instance = Asset.objects.filter(id=pk).first()
-
-        platform = None
-        if self.instance:
-            platform = self.instance.platform
-        elif request.query_params.get('platform'):
-            platform_id = request.query_params.get('platform')
-            platform_id = int(platform_id) if platform_id.isdigit() else 0
-            platform = Platform.objects.filter(id=platform_id).first()
-
-        if not platform:
-            return default_field
-        custom_fields = platform.custom_fields
-        if not custom_fields:
-            return default_field
-        name = platform.name.title() + 'CustomSerializer'
-        return create_serializer_class(name, custom_fields)()
 
     @staticmethod
     def perform_nodes_display_create(instance, nodes_display):
@@ -336,8 +303,8 @@ class DetailMixin(serializers.Serializer):
 
     def get_instance(self):
         request = self.context.get('request')
-        if not self.instance and uuid_pattern.findall(request.path):
-            pk = uuid_pattern.findall(request.path)[0]
+        if not self.instance and UUID_PATTERN.findall(request.path):
+            pk = UUID_PATTERN.findall(request.path)[0]
             self.instance = Asset.objects.filter(id=pk).first()
         return self.instance
